Make IAM policy generation more testable
