+++ src/JustSaying/AwsTools/MessageHandling/SnsPolicy.cs
@@ -1,5 +1,3 @@
-using System.Text.Json;
-using System.Text.RegularExpressions;
 using Amazon.SimpleNotificationService;
 using Amazon.SimpleNotificationService.Model;
 
@@ -9,52 +7,8 @@ internal static class SnsPolicy
 {
     internal static async Task SaveAsync(SnsPolicyDetails policyDetails, IAmazonSimpleNotificationService client)
     {
-        var sourceAccountId = ExtractSourceAccountId(policyDetails.SourceArn);
-        var policyJson = $@"{{
-    ""Version"" : ""2012-10-17"",
-    ""Statement"" : [
-        {{
-            ""Sid"" : ""{Guid.NewGuid().ToString().Replace("-", "")}"",
-            ""Effect"" : ""Allow"",
-            ""Principal"" : {{
-                ""AWS"" : ""*""
-            }},
-            ""Action""    : [
-                ""sns:GetTopicAttributes"",
-                ""sns:SetTopicAttributes"",
-                ""sns:AddPermission"",
-                ""sns:RemovePermission"",
-                ""sns:DeleteTopic"",
-                ""sns:Subscribe"",
-                ""sns:Publish""
-            ],
-            ""Resource""  : ""{policyDetails.SourceArn}"",
-            ""Condition"" : {{
-                ""StringEquals"" : {{
-                    ""AWS:SourceOwner"" : ""{sourceAccountId}""
-                }}
-            }}
-        }},
-        {{
-            ""Sid"" : ""{Guid.NewGuid().ToString().Replace("-", "")}"",
-            ""Effect"" : ""Allow"",
-            ""Principal"" : {{
-                ""AWS"" : {JsonSerializer.Serialize(policyDetails.AccountIds)}
-            }},
-            ""Action""    : ""sns:Subscribe"",
-            ""Resource""  : ""{policyDetails.SourceArn}""
-        }}
-    ]
-}}";
+        var policyJson = SnsPolicyBuilder.BuildPolicyJson(policyDetails);
         var setQueueAttributesRequest = new SetTopicAttributesRequest(policyDetails.SourceArn, "Policy", policyJson);
-
         await client.SetTopicAttributesAsync(setQueueAttributesRequest).ConfigureAwait(false);
-    }
-
-    private static string ExtractSourceAccountId(string sourceArn)
-    {
-        //Sns Arn pattern: arn:aws:sns:region:account-id:topic
-        var match = Regex.Match(sourceArn, "(.*?):(.*?):(.*?):(.*?):(.*?):(.*?)", RegexOptions.None, Regex.InfiniteMatchTimeout);
-        return match.Groups[5].Value;
     }
 }
