+++ tests/JustSaying.UnitTests/AwsTools/MessageHandling/Sns/Policy/Approvals/SnsPolicyBuilderTests.WhenGeneratingPolicy_ForSnsTopic_ThenTheApprovedJsonShouldBeReturned.approved.txt
@@ -1,0 +1,36 @@
+{
+    "Version" : "2012-10-17",
+    "Statement" : [
+        {
+            "Sid" : "<sid1>",
+            "Effect" : "Allow",
+            "Principal" : {
+                "AWS" : "*"
+            },
+            "Action"    : [
+                "sns:GetTopicAttributes",
+                "sns:SetTopicAttributes",
+                "sns:AddPermission",
+                "sns:RemovePermission",
+                "sns:DeleteTopic",
+                "sns:Subscribe",
+                "sns:Publish"
+            ],
+            "Resource"  : "arn:aws:sns:ap-southeast-2:123456789012:topic",
+            "Condition" : {
+                "StringEquals" : {
+                    "AWS:SourceOwner" : "123456789012"
+                }
+            }
+        },
+        {
+            "Sid" : "<sid2>",
+            "Effect" : "Allow",
+            "Principal" : {
+                "AWS" : null
+            },
+            "Action"    : "sns:Subscribe",
+            "Resource"  : "arn:aws:sns:ap-southeast-2:123456789012:topic"
+        }
+    ]
+}
